[PATCH] transmeta efficeon support and cpuid update
This patch adds efficeon as a cpu option, and makes a small update to the
transmeta cpuid code.  (i wasn't sure if the various doc files are UTF-8...
if they are, then the e should be a U-275 ;)

The compile options may not be ideal, but they're probably close.  i used
-march=pentium3, but -march=pentium4 would have been good enough too.

The cpuid update teaches transmeta.c about the extended processor revision
present in cpuid level 0x80860002...  the external documentation does not
indicate how to break apart this field, and instructs only that the 32-bit
value should be printed in hex (alas).

Signed-off-by: dean gaudet <dean@arctic.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41768728WDrzEUqVNeWs2RhCAnHToQ