[PATCH] reiserfs: support for REISERFS_UNSUPPORTED_OPT notation
This patch adds a REISERFS_UNSUPPORTED_OPT flag to denote when a mount
option is allowable, but is unsupported in the running configuration.  This
allows the potential for the set of mount options to be consistent,
regardless of what features the kernel is compiled with.

Rather than failing the mount, a warning is issued and the mount succeeds.

Signed-off-by: Jeff Mahoney <jeffm@novell.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41768780AlkFyZUzCVWUXjZOs-e6uw