[NET]: Add new sysfs attribute 'carrier' for net devices.
From: Jesper Juhl <juhl-lkml@dif.dk>

The patch adds a new sysfs attribute called carrier for net devices that
exposes the result of netif_carrier_ok().  This allows a user to quickly
and easily determine if a given netdevice has carrier or not.  Reading
/sys/class/net/<interface>/carrier yields a '1' when there is a carrier and
a '0' when there is not.

Signed-off-by: Jesper Juhl <juhl-lkml@dif.dk>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

BKrev: 417892de2kxWzgpko-p3p9tw-6RVHw