Merge pobox.com:/garz/repo/netdev-2.6/acenic into pobox.com:/garz/repo/net-drivers-2.6
2004/10/06 23:44:56-04:00 akpm
[PATCH] fix acenic hotplug

From: Anton Blanchard <anton@samba.org>

Now that the acenic has been converted to the hotplug API a number of uses
of __init and __initdata have to be changed to __devinit and __devinitdata.

Acked-by: Jes Sorensen <jes@trained-monkey.org>
Signed-off-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>

2004/10/06 23:39:55-04:00 jgarzik
Merge pobox.com:/spare/repo/linux-2.6
into pobox.com:/spare/repo/netdev-2.6/acenic

2004/09/20 14:03:45-04:00 shemminger
[PATCH] (4/4) acenic - don't spin forever in hard_start_xmit

If driver is stuck due to ring full and hardware or link error, don't spin
forever in the hard_start_xmit routine.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>

2004/09/20 14:03:34-04:00 shemminger
[PATCH] (3/4) acenic - __iomem warnings cleanup

This cleans all the compile warnings and most of the sparse warnings
for the acenic driver relating to io memory space.
Remaining warnings are because tx_ring can be either in i/o or not
depending on the version of the card.
Not tested on old TIGON card.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>

2004/09/20 14:03:22-04:00 shemminger
[PATCH] (2/4) acenic - eliminate MAX_SKB_FRAGS #if

Since MAX_SKB_FRAGS is defined in both 2.4 and 2.6, it makes sense
to eliminate this old #if code.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>

2004/09/20 14:03:11-04:00 shemminger
[PATCH] (1/4) acenic - use netdev_priv

Trivial, use netdev_priv

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>

BKrev: 417caba8bDLlMOKc8GSscVAp38wnaw