[PATCH] ppc64: create iommu_free_table()
This patch is from John Rose <johnrose@austin.ibm.com>.

The patch below creates iommu_free_table().  Iommu tables are not currently
freed in PPC64.  This could cause a memory leak for DLPAR of an EADS slot.
The function verifies that there are no outstanding TCE entries for the
range of the table before freeing it.  I added a call to iommu_free_table()
to the code that dynamically removes a device node.  This should be fairly
symmetrical with the table allocation, which happens during dynamic
addition of a device node.

Signed-off-by: John Rose <johnrose@austin.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 417da35btOQ0XJP0DF3YLR2q3olJHw