[PATCH] md: remove md_flush_all
Following are 7 patches for md

They all grew out of a desire to redo the locking in unplug_slave.  Getting
and dropping a spinlock so often for very little gain (it would be nearly
impossible to lose the relevant race) really bothered me.

I finally figured that I could reply it with rcu locking which is very light
wait, and quite up to the task.

One the way I found an number of inconsistencies that needed cleaning up and
even a few bugs to fix.  The first 6 patches deal with these inconsistencies
and bugs.  The last redoes the locking for adding/removing/accessing devices
within md personalities.


This patch:

md_flush_all() isn't needed as each personality defines its own
issue_flush_fn.

Signed-off-by: Neil Brown <neilb@cse.unsw.edu.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 417da43aOnCzJNxngptADSmcvuuRRw