[PATCH] Weak symbols in modules and versioned symbols
Weak symbol may be unavailable in kernel, but if it is available, its
signature should be same as was at the build time.  If we do not attach
signatures to weak symbols, kernel is tainted when such module is loaded.

vmmon: no version for "sys_ioctl" found: kernel tainted.

I also believe that it is safer to add & check signatures here - module
wants either sys_ioctl with right signature, or no sys_ioctl at all, not
sys_ioctl with different signature (which signals that there is some
misbuild occuring).

Signed-off-by: Petr Vandrovec <vandrove@vc.cvut.cz>
Signed-off-by: Rusty Russell <rusty@rustcorp.com.au> (forwarded)
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 417da4e7TI_nzDgaAEXFWTg32Rg_wg