[PATCH] idr_remove safety checking
idr_remove() should fail gracefully and warn if the id being removed is not
valid.

The attached patch should do the job without additional overhead.

With the existing code, removing an id which was not allocated could remove
a valid id which shares the same lowest layer of the radix tree.

I ran a kernel with this patch but have not done any tests to force
a failure.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 417da70bJtmQzAILuEcTo3Cjy7d_FQ