[PATCH] ppc64: Add new "Maple" platform support
This adds support for the Maple 970FX Eval Board.  It adds the basic
arch support.  For the Maple to be fully functional, it needs a couple
more patches to be applied for IDE and Ethernet that are currently
pending with the respective maintainers of those subsystems.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 417e64eegVwTnjngcH0IDZ6pVhv_0w