[PATCH] omit CommitAvail
CommitLimit was a good addition to /proc/meminfo, but we don't usually show
both what's used and what's free: don't waste lines of screenspace, omit
CommitAvail, let the user do the arithmetic as with all the others.  And in
updating that Documentation, removed the long-gone ReverseMaps.

Signed-off-by: Hugh Dickins <hugh@veritas.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41810d9bP6cL3LqZxxm0ufYbjtcNkA