[PATCH] PPC/PPC64: Fix FP state corruption on UP
Unfortunately the patch Ben sent last week to fix a bug in the saving and
restoring of floating-point and altivec context across signal handlers
introduced another bug, which tends to corrupt the FP and altivec contexts
of other tasks.  This patch fixes the problem for both ppc32 and ppc64.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41810e0blGOjUVLDcEji8lvJHPItxA