[PATCH] RCU: rcu_assign_pointer() removal of memory barriers
This patch adds the rcu_assign_pointer() API that helps reduce the need for
explicit memory barriers in code that uses RCU.  This API buries the required
memory barriers in a macro that also does the assignment.  This has been
tested successfully on i386 and ppc64.

Signed-off-by: <paulmck@us.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4181100dE6e3XhBCRxICKr0hHe-4ag