Merge http://lia64.bkbits.net/linux-ia64-release-2.6.10 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/10/27 11:36:56-07:00 tony.luck
[IA64] fix(?) unwind data for ia64_monarch_init_handler

Roland> I can see the ERROR: comes from unwcheck.pl but I have no
Roland> idea what it is checking (something to do with unwind info?)
Roland> let alone how to fix the problem.

Roland> Is this benign or is there something to be fixed?

David> Both: it needs to be fixed, but it's (mostly) benign since the only
David> time we're executing that code is when the unwinder can't run anyhow,
David> since we're in the middle of delivering an MCA.  That may (and
David> probably will) change in the future, so it's something that should be
David> fixed.

David> I was hoping someone who knows and could test the MCA code better
David> would fix it but that hasn't happened yet, so it's time for plan B.

David> Tony, below is a patch that fixes the unwind info enough that the
David> warning/error goes away.  The unwind-info may not be 100% correct
David> though.  Someone who can test/exercise this code may want to check on
David> that.

Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/27 17:07:16+00:00 steiner
[IA64-SGI] Update asm-ia64/sn/sn_cpuid.h macros

(I missed a file in the cleanup code that I sent yesterday.)

SGI SN code currently makes assumptions about the bits in the LID
register. These assumptions do not conform to the bit specifications
from Intel. For example, SN currently assumes that bits [28:16] of
the LID contain the physical node ID of a node.

This patch eliminates these assumptions. A SAL call is now used to translate
LID values to the NASID/subnode/slice values that are needed for SN
platforms. The results of the SAL call are saved in the SN nodepda.

Signed-off-by: Jack Steiner <steiner@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/27 16:21:44+00:00 jbarnes
[IA64] fix machine vectors for mmiowb

I left out some of the necessary machine vector magic in the mmiowb patch
which caused the generic build to break.  This patch fixes it up by properly
#defining it and #undefing it and adding a routine to lib/io.c.

Signed-off-by: Jesse Barnes <jbarnes@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/27 03:40:21+00:00 tony.luck
Merge intel.com:/data/home/aegl/BK/Linus
into intel.com:/data/home/aegl/BK/linux-ia64-release-2.6.10

2004/10/26 23:50:29+00:00 tony.luck
[IA64] Need <asm/meminit.h> for GRANULEROUNDDOWN

Alex's change for mem=/max_addr= now needs to include meminit.h

Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 23:50:13+00:00 pfg
[IA64-SGI] only allocate irq if the device can interrupt

Signed-off-by: Patrick Gefre <pfg@sgi.com>
Acked-by: Mike Habeck <habeck@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 23:49:56+00:00 steiner
[IA64-SGI] Delete simulator support from SN idle loop

Delete hack for supporting simulator for SN platforms. This capability
has been moved into the simulator environment & is no longer required
in the kernel.

Signed-off-by: Jack Steiner <steiner@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 23:49:38+00:00 steiner
[IA64-SGI] Update asm-ia64/sn/sn_cpuid.h macros

SGI SN code currently makes assumptions about the bits in the LID
register. These assumptions do not conform to the bit specifications
from Intel. For example, SN currently assumes that bits [28:16] of
the LID contain the physical node ID of a node.

This patch eliminates these assumptions. A SAL call is now used to translate
LID values to the NASID/subnode/slice values that are needed for SN
platforms. The results of the SAL call are saved in the SN nodepda.

Signed-off-by: Jack Steiner <steiner@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 20:37:32+00:00 steiner
[IA64] Delete obsolete code from SGI console driver

Delete obsolete code that supports the SGI simulator. This
support code has been moved into the fakeprom and is no
longer required in the kernel.

Signed-off-by: Jack Steiner <steiner@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 20:33:21+00:00 jasonuhl
[IA64] fix pgtable.h comments

Remove some statements from comments in <asm-ia64/pgtable.h> whose
correctness is page size dependent ... and based on using an 8k
page size, which almost nobody actually uses.

Signed-off-by: Jason Uhlenkott <jasonuhl@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 19:24:59+00:00 kaos
[IA64] Correct references from text/data to init.text/data.

These errors were found by 'make buildcheck'.

Signed-off-by: Keith Owens <kaos@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 19:14:39+00:00 markgw
[IA64] sn_hwperf correctly handle bricks with multiple slabs

The procfs handler for /proc/sgi_sn/sn_topology did
not correctly handle multiple slabs in the same brick,
e.g. a brick containing a compute node and an ionode.

Signed-off-by: Mark Goodwin <markgw@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 19:03:51+00:00 kaos
[IA64] Correct bit test for salinfo oem decode

The valid.oem_data bit is in different positions in each section.  Make
the test for valid oem data a section specific test.

Signed-off-by: Keith Owens <kaos@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 18:41:39+00:00 eranian
[IA64] misc small patches for perfmon

change-log:
- update comment for pfm_do_fasync()
- fix return value for pfm_smpl_buffer_alloc(). Now
  return ENOMEM instead of EAGAIN when buffer too big
  for RLIMIT_MEMLOCK.
- added missing vm_pgoff initialization in pfm_smpl_buffer_alloc()
- added flags tro debug print in pfm_write_pmcs
- added seed and mask to debug print in pfm_write_pmds
- shorten some of the debug prints
- remove bogus sanity check from pfm_save_regs() in SMP.
  this could cause invalid PMU state restore

signed-off-by: eranian@hpl.hp.com
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 18:36:52+00:00 alex.williamson
[IA64] efi.c: fix mem= & max_addr=

With this change, there's some extra fuzz introduced that a max_addr
specification will get rounded down to a granule boundary and memory
quantity, when using mem=, will be within a granule size of the
requested amount.  Let me know if anyone finds more problems with it.

Signed-off-by: Alex Williamson <alex.williamson@hp.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/10/26 18:25:01+00:00 tony.luck
Merge intel.com:/data/home/aegl/BK/Linus
into intel.com:/data/home/aegl/BK/linux-ia64-release-2.6.10

2004/10/26 18:23:18+00:00 akpm
[IA64] Need <asm/uaccess.h> for KERNEL_DS & set_fs() definitions.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Tony Luck <tony.luck@intel.com>

BKrev: 41814444dQfyPQ1pk04ANmN42jlXKw