[PATCH] therm_adt746x: gradually change fan speed and start both fans on Albooks
Here's the patch that supercedes the one I asked you to drop.  It changes
the fan speed logic so that the fans will go incrementally fast instead of
using just two steps.  Also, it starts and stops both fans instead of just
one on Powerbooks 15" and 17" (functionality asked by most of the Albooks
users on Debian's ML, and justified by the fact that fan 1 and fan 2 do not
correspond to CPU and GPU, as we previously thought).

Signed-off-by: Colin Leroy <colin@colino.net>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 418814c7rgN6CyFi4gGL6xs5V0Kjeg