Merge bk://kernel.bkbits.net/davem/net-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/11/03 16:04:55-08:00 Andries.Brouwer
[PATCH] avoid semi-infinite loop when mounting bad ext2

The routine ext2_readdir() will, when reading a directory page
returns an error, try the next page, without reporting the
error to user space. That is bad, and the patch below changes that.

In my case the filesystem was damaged, and ext2_readdir wanted
to read 60000+ pages and wrote as many error messages to syslog
("attempt to access beyond end"), not what one wants.

[no doubt a similar patch is appropriate for ext3]

BKrev: 41897aaelM1ttTzaBuv9sPAk4u8MZg